Rename toMultiVar.vim
I realise that naming is contentious. To avoid as much argument as
possible I am using the naming convention from this:

    https://google.github.io/styleguide/vimscriptguide.xml?showone=Naming#Naming

The files in this plugin fall under the augroup_names_like_this rule.
This is because you are defining the entire functionality of your plugin
using an autoload group.